Tests/LibWeb/Text/expected/css/insert-import-rule-into-constructed-stylesheet.txt

Inserting an @import rule into a constructed stylesheet throws: SyntaxError
